requirements.txt
numpy
scipy
tqdm
matplotlib

open3d
opencv-python

torch-scatter

jupyter
imageio

glog
icecream

pandas
pyrealsense2

pybind11
gdown

colored_glog

kornia==0.5.3
cupy-cuda117
timm
omegaconf
pytorch_msssim
tensorboard
imageio[ffmpeg]
lpips